module structured upon coding standards
