+++ package.json
@@ -2,7 +2,7 @@
     "name": "kitsune-br-gzip-compression-lib",
     "version": "1.0.0",
     "description": "Kitsune brotli and gzip compression utils library for nodejs",
-    "main": "compression.js",
+    "main": "index.js",
     "type": "module",
     "dependencies": {
         "brotli": "^1.3.2",
@@ -15,7 +15,7 @@
     },
     "devDependencies": {},
     "scripts": {
-        "start": "node compression.js"
+        "start": "node index.js"
     },
     "repository": {
         "type": "git",
